+++ tests/test_managers.py
@@ -136,7 +136,9 @@ class TestResourceManager(unittest.TestCase):
 
     def test_create_empty_resource(self):
         project = self.redmine.project.new()
-        self.assertEqual(project.attributes, project._relations)
+        defaults = dict((include, None) for include in project._includes)
+        defaults.update(dict((relation, None) for relation in project._relations))
+        self.assertEqual(project.attributes, defaults)
 
     @mock.patch('requests.put')
     def test_update_resource(self, mock_put):